+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## v0.3.0 (not released)
 
 - change (breaking!): entries must know be written as `.exs` files. Otherwise they will be ignored.
-- improvement: storybook is now fully responsive
+- change (breaking!): `variations` have been rebranded as `stories`.
+- improvement: storybook is now fully responsive.
 - improvement: (stateless) component entries no longer require to define a `component/0` function.
 
 ## v0.2.0 (2022-07-30)
@@ -11,7 +12,7 @@
 - feature: new tab to browse your component sources
 - feature: work-in-progress component documentation source
 - feature: new page entry support, which allows you create custom pages within your storybook
-- improvement: introduced `%VariationGroup{}` to render mulitple variations in a single page div.
+- improvement: introduced `%StoryGroup{}` to render mulitple stories in a single page div.
 
 ## v0.1.0 (2022-07-26)